+++ src/components/AbsoluteCard/AbsoluteCard.tsx
@@ -1,8 +1,8 @@
 import { ReactNode } from "react";
 
 import cn from "@/utils/cn";
-import AbsoluteCardImage from "@/components/AbsoluteCard/AbsoluteCardImage";
-import AbsoluteCardContent from "@/components/AbsoluteCard/AbsoluteCardContent";
+import { AbsoluteCardImage } from "@/components/AbsoluteCard/AbsoluteCardImage";
+import { AbsoluteCardContent } from "@/components/AbsoluteCard/AbsoluteCardContent";
 
 type AbsoluteCardProps = {
   className?: string;
