+++ src/components/Testimonial/components/TestimonialRating.tsx
@@ -4,7 +4,7 @@ type TestimonialRatingProps = {
   rating: number;
 };
 
-export default function TestimonialRating({ rating }: TestimonialRatingProps) {
+export function TestimonialRating({ rating }: TestimonialRatingProps) {
   return (
     <div className="flex">
       {Array.from({ length: 5 }, (_, ind) => {
