--- src/features/Slider/index.ts
@@ -1,2 +1,0 @@
-export type { Slide } from "@/features/Slider/types/types";
-export { default as Slider } from "@/features/Slider/components/Slider";
