--- src/features/Slider/types/types.tsx
@@ -1,10 +1,0 @@
-export type Slide = {
-  image: {
-    src: string;
-  };
-  content: {
-    title: string;
-    subtitle: string;
-    category: string;
-  };
-};
