+++ src/layouts/Header/components/HeaderActions.tsx
@@ -3,13 +3,13 @@ import React from "react";
 import cn from "@/utils/cn";
 import Button from "@/components/Button";
 import { CartIcon, HeartIcon } from "@/components/Icons";
-import HeaderAuthBtn from "@/layouts/Header/components/HeaderAuthBtn";
+import { HeaderAuthBtn } from "@/layouts/Header/components/HeaderAuthBtn";
 
 type HeaderActionsProps = {
   className?: string;
 };
 
-export default function HeaderActions({ className }: HeaderActionsProps) {
+export function HeaderActions({ className }: HeaderActionsProps) {
   return (
     <div className={cn("[&_button]:px-4", className)}>
       <Button className="group">
