+++ src/layouts/Header/components/HeaderAuthBtn.tsx
@@ -1,12 +1,12 @@
 "use client";
 
 import React from "react";
+import { signIn, signOut, useSession } from "next-auth/react";
 
 import Button from "@/components/Button";
 import { LogoutIcon, UserIcon } from "@/components/Icons";
-import { signIn, signOut, useSession } from "next-auth/react";
 
-export default function HeaderAuthBtn() {
+export function HeaderAuthBtn() {
   const { data } = useSession();
 
   function authOnClick() {
