+++ src/layouts/Header/components/HeaderBurger.tsx
@@ -6,7 +6,7 @@ import { Navbar } from "@/layouts/Sidebar";
 import OverlayPortal from "@/components/Overlay";
 import { HeaderActions, HeaderSearch } from "@/layouts/Header";
 
-export default function HeaderBurger() {
+export function HeaderBurger() {
   const [isOpen, setIsOpen] = useState(false);
 
   return (
