+++ src/layouts/Header/components/HeaderSearch.tsx
@@ -1,6 +1,7 @@
-import { SearchIcon } from "@/components/Icons";
 import React from "react";
+
 import cn from "@/utils/cn";
+import { SearchIcon } from "@/components/Icons";
 
 type HeaderSearchProps = {
   className?: string;
