+++ src/layouts/Header/index.tsx
@@ -1,4 +1,4 @@
 export { default as Header } from "@/layouts/Header/components/Header";
-export { default as HeaderBurger } from "@/layouts/Header/components/HeaderBurger";
+export { HeaderBurger } from "@/layouts/Header/components/HeaderBurger";
+export { HeaderActions } from "@/layouts/Header/components/HeaderActions";
 export { default as HeaderSearch } from "@/layouts/Header/components/HeaderSearch";
-export { default as HeaderActions } from "@/layouts/Header/components/HeaderActions";
